requirements.txt
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
flask-migrate==2.5.3
flask-script==2.0.6
python-dotenv
psycopg2
gunicorn
